pyproject.toml
[tool.poetry]
name = "tdjango"
version = "0.1.0"
description = "koka"
authors = ["Hardeep Kumar <hardeepkumar1815@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
yapf = "^0.31.0"
Pillow = "^8.3.1"
django-ckeditor = "^6.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
